requirements.txt
PyQt6>=6.4.0
python-escpos>=3.0
PyInstaller>=6.0